requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
requests==2.31.0
protobuf==4.25.2
gtfs-realtime-bindings==1.0.0
sqlalchemy==1.4.51
psycopg2-binary==2.9.10
python-dotenv==1.0.0
apscheduler==3.10.4
websockets==11.0.3
redis==5.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
aiohttp==3.11.13 